# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FastFreela API",
        "description": "Documentação da API do FastFreela",
        "contact": {
            "email": "pedroa990@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Lista todas as categorias",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "responses": {
                    "200": {
                        "description": "Lista de categorias"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Cria nova categoria",
                "operationId": "bb02f0320882db37f99574196577e446",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Categoria X"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Exibe dados de uma categoria",
                "operationId": "4cdf41756315109f1efc72787db5830a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da categoria"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualiza uma categoria",
                "operationId": "19954b1f8aa4916b5f032dc28ea36a24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Categoria X atualizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Remove uma categoria",
                "operationId": "f1b8af6e84bd52b9594d98bde748ebd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria deletada com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/services": {
            "get": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Lista todos os serviços",
                "operationId": "0f76dda9be6628c47b978d3b7077529a",
                "responses": {
                    "200": {
                        "description": "Lista de serviços"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Cria novo serviço",
                "operationId": "ee694c210db03e85b35b350b82c18f1b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Novo serviço"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição do serviço"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serviço criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Exibe dados de um serviço",
                "operationId": "ee0c95755e1374f6e92ee6ae1cdddfb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do serviço"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Atualiza um serviço",
                "operationId": "b9a49c8480cb32ec2fa69841690d863e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Serviço atualizado"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Nova descrição"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Serviço atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Remove um serviço",
                "operationId": "cbb6bc64cf61a4812d1c6da5421fd545",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Serviço deletado com sucesso"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
...
    },
...
}